feat(web): as métricas completas da sessão viram aba no aside da direita
O modal com tudo o que a store sabe da conversa só existia nas listas do
dashboard. Dentro do workspace de sessões — onde a pergunta realmente
aparece — não havia como chegar nele, e o card de métricas do topo é o
que um popover de 300px consegue mostrar e nada além.

Agora o card termina num link ("ver tudo no painel") que abre a aba
**Métricas** no aside da direita, onde os números ficam ao lado da
conversa de que falam.

- `SessionDrilldown.tsx` é o HEAD (o que a sessão é) e o BODY (o que ela
  gastou), usados pelas duas superfícies. Uma segunda cópia seria um
  segundo conjunto de respostas sobre a mesma sessão.
- O modal continua existindo para as três telas que não têm aside — o
  board "quem lidera" da Home, a aba Actions do repositório e os
  componentes de layout personalizado. Removê-lo deixaria essas linhas
  inertes.
- **O layout segue a COLUNA, não a viewport.** As grades densas eram
  decididas por `useIsMobile()`, o que está certo para um diálogo que é
  980px ou o telefone inteiro e errado para um painel que o leitor
  arrasta entre 280 e 900px. O body se MEDE (`NARROW_AT`), e `isMobile`
  responde só pelo frame anterior à medição — `useLayoutEffect`, então
  ninguém vê a grade larga dentro de um painel estreito.
- A aba é ABSENTE quando a store não tem registro da conversa, e o link
  é retido pelo mesmo fato: um link para uma aba que não existe é o
  controle morto que este produto recusa em todo lugar. A aba também
  passa na frente da recusa de leitura do transcript, como a de tarefa —
  são fontes diferentes.
- `count` das abas passa a ter três estados: `null` "ninguém perguntou
  ainda", um número, e AUSENTE para uma aba que não conta nada. Sem
  isso a grade imprimiria `—` sob "abra a aba para contar", promessa que
  esta aba nunca poderia cumprir.

Verificado a 1600px (aside 620px, layout largo) e a 390px (aside em tela
cheia, layout estreito, alvo do link 274x44), com o modal ainda abrindo
nos dois tamanhos e `scrollWidth === innerWidth`.
